Move DoubleTests
