+++ Tests/Runner/Batch4/Bridge.Test/bridge.clienttest.batch4.runner.js
@@ -114,32 +114,6 @@ Bridge.assembly("Bridge.Test.Bridge.ClientTest.Batch4", function ($asm, globals)
             QUnit.test("DecimalTests - ParseWorks_SPI_1586", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DecimalTests.ParseWorks_SPI_1586);
             QUnit.test("DecimalTests - TryParseWorks_SPI_1586", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DecimalTests.TryParseWorks_SPI_1586);
             QUnit.test("DecimalTests - ImplementationTests_SPI_1588_1590_1650", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DecimalTests.ImplementationTests_SPI_1588_1590_1650);
-            QUnit.test("DoubleTests - TypePropertiesAreCorrect_SPI_1717", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DoubleTests.TypePropertiesAreCorrect_SPI_1717);
-            QUnit.test("DoubleTests - DefaultValueIs0", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DoubleTests.DefaultValueIs0);
-            QUnit.test("DoubleTests - CreatingInstanceReturnsZero", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DoubleTests.CreatingInstanceReturnsZero);
-            QUnit.test("DoubleTests - ConstantsWork", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DoubleTests.ConstantsWork);
-            QUnit.test("DoubleTests - DefaultConstructorReturnsZero", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DoubleTests.DefaultConstructorReturnsZero);
-            QUnit.test("DoubleTests - FormatWorks", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DoubleTests.FormatWorks);
-            QUnit.test("DoubleTests - ToStringWithFormatWorks", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DoubleTests.ToStringWithFormatWorks);
-            QUnit.test("DoubleTests - ToStringWithFormatAndProviderWorks", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DoubleTests.ToStringWithFormatAndProviderWorks);
-            QUnit.test("DoubleTests - IFormattableToStringWorks", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DoubleTests.IFormattableToStringWorks);
-            QUnit.test("DoubleTests - ToStringWorks", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DoubleTests.ToStringWorks);
-            QUnit.test("DoubleTests - ToExponentialWorks", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DoubleTests.ToExponentialWorks);
-            QUnit.test("DoubleTests - ToExponentialWithFractionalDigitsWorks", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DoubleTests.ToExponentialWithFractionalDigitsWorks);
-            QUnit.test("DoubleTests - ToFixed", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DoubleTests.ToFixed);
-            QUnit.test("DoubleTests - ToFixedWithFractionalDigitsWorks", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DoubleTests.ToFixedWithFractionalDigitsWorks);
-            QUnit.test("DoubleTests - ToPrecisionWorks", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DoubleTests.ToPrecisionWorks);
-            QUnit.test("DoubleTests - ToPrecisionWithPrecisionWorks", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DoubleTests.ToPrecisionWithPrecisionWorks);
-            QUnit.test("DoubleTests - IsPositiveInfinityWorks_SPI_1600", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DoubleTests.IsPositiveInfinityWorks_SPI_1600);
-            QUnit.test("DoubleTests - IsNegativeInfinityWorks", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DoubleTests.IsNegativeInfinityWorks);
-            QUnit.test("DoubleTests - IsInfinityWorks", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DoubleTests.IsInfinityWorks);
-            QUnit.test("DoubleTests - IsFiniteWorks", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DoubleTests.IsFiniteWorks);
-            QUnit.test("DoubleTests - IsNaNWorks", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DoubleTests.IsNaNWorks);
-            QUnit.test("DoubleTests - GetHashCodeWorks", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DoubleTests.GetHashCodeWorks);
-            QUnit.test("DoubleTests - EqualsWorks", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DoubleTests.EqualsWorks);
-            QUnit.test("DoubleTests - IEquatableEqualsWorks", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DoubleTests.IEquatableEqualsWorks);
-            QUnit.test("DoubleTests - CompareToWorks", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DoubleTests.CompareToWorks);
-            QUnit.test("DoubleTests - IComparableCompareToWorks", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DoubleTests.IComparableCompareToWorks);
             QUnit.test("Int16Tests - TypePropertiesAreCorrect_SPI_1717", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.Int16Tests.TypePropertiesAreCorrect_SPI_1717);
             QUnit.test("Int16Tests - CastsWork", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.Int16Tests.CastsWork);
             QUnit.test("Int16Tests - DefaultValueIs0", Bridge.Test.Runtime.BridgeClientTestBatch4Runner.Int16Tests.DefaultValueIs0);
@@ -1381,156 +1355,6 @@ Bridge.assembly("Bridge.Test.Bridge.ClientTest.Batch4", function ($asm, globals)
                 var $t;
                 if (this.context == null) {
                     this.context = ($t = new Bridge.Test.Runtime.FixtureContext(), $t.Project = "Batch4", $t.ClassName = "Bridge.ClientTest.Batch4.DelegateTests", $t.File = "Batch4\\DelegateTests.cs", $t);
-                }
-                return this.context;
-            }
-        }
-    });
-
-    Bridge.define("Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DoubleTests", {
-        inherits: [Bridge.Test.Runtime.TestFixture$1(Bridge.ClientTest.Batch4.SimpleTypes.DoubleTests)],
-        statics: {
-            methods: {
-                TypePropertiesAreCorrect_SPI_1717: function (assert) {
-                    var $t;
-                    var t = Bridge.Test.Runtime.TestFixture$1(Bridge.ClientTest.Batch4.SimpleTypes.DoubleTests).BeforeTest(true, assert, Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DoubleTests, void 0, ($t = new Bridge.Test.Runtime.TestContext(), $t.Method = "TypePropertiesAreCorrect_SPI_1717()", $t.Line = "10", $t));
-                    t.Fixture.TypePropertiesAreCorrect_SPI_1717();
-                },
-                DefaultValueIs0: function (assert) {
-                    var $t;
-                    var t = Bridge.Test.Runtime.TestFixture$1(Bridge.ClientTest.Batch4.SimpleTypes.DoubleTests).BeforeTest(true, assert, Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DoubleTests, void 0, ($t = new Bridge.Test.Runtime.TestContext(), $t.Method = "DefaultValueIs0()", $t.Line = "37", $t));
-                    t.Fixture.DefaultValueIs0();
-                },
-                CreatingInstanceReturnsZero: function (assert) {
-                    var $t;
-                    var t = Bridge.Test.Runtime.TestFixture$1(Bridge.ClientTest.Batch4.SimpleTypes.DoubleTests).BeforeTest(true, assert, Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DoubleTests, void 0, ($t = new Bridge.Test.Runtime.TestContext(), $t.Method = "CreatingInstanceReturnsZero()", $t.Line = "43", $t));
-                    t.Fixture.CreatingInstanceReturnsZero();
-                },
-                ConstantsWork: function (assert) {
-                    var $t;
-                    var t = Bridge.Test.Runtime.TestFixture$1(Bridge.ClientTest.Batch4.SimpleTypes.DoubleTests).BeforeTest(true, assert, Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DoubleTests, void 0, ($t = new Bridge.Test.Runtime.TestContext(), $t.Method = "ConstantsWork()", $t.Line = "49", $t));
-                    t.Fixture.ConstantsWork();
-                },
-                DefaultConstructorReturnsZero: function (assert) {
-                    var $t;
-                    var t = Bridge.Test.Runtime.TestFixture$1(Bridge.ClientTest.Batch4.SimpleTypes.DoubleTests).BeforeTest(true, assert, Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DoubleTests, void 0, ($t = new Bridge.Test.Runtime.TestContext(), $t.Method = "DefaultConstructorReturnsZero()", $t.Line = "63", $t));
-                    t.Fixture.DefaultConstructorReturnsZero();
-                },
-                FormatWorks: function (assert) {
-                    var $t;
-                    var t = Bridge.Test.Runtime.TestFixture$1(Bridge.ClientTest.Batch4.SimpleTypes.DoubleTests).BeforeTest(true, assert, Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DoubleTests, void 0, ($t = new Bridge.Test.Runtime.TestContext(), $t.Method = "FormatWorks()", $t.Line = "69", $t));
-                    t.Fixture.FormatWorks();
-                },
-                ToStringWithFormatWorks: function (assert) {
-                    var $t;
-                    var t = Bridge.Test.Runtime.TestFixture$1(Bridge.ClientTest.Batch4.SimpleTypes.DoubleTests).BeforeTest(true, assert, Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DoubleTests, void 0, ($t = new Bridge.Test.Runtime.TestContext(), $t.Method = "ToStringWithFormatWorks()", $t.Line = "75", $t));
-                    t.Fixture.ToStringWithFormatWorks();
-                },
-                ToStringWithFormatAndProviderWorks: function (assert) {
-                    var $t;
-                    var t = Bridge.Test.Runtime.TestFixture$1(Bridge.ClientTest.Batch4.SimpleTypes.DoubleTests).BeforeTest(true, assert, Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DoubleTests, void 0, ($t = new Bridge.Test.Runtime.TestContext(), $t.Method = "ToStringWithFormatAndProviderWorks()", $t.Line = "81", $t));
-                    t.Fixture.ToStringWithFormatAndProviderWorks();
-                },
-                IFormattableToStringWorks: function (assert) {
-                    var $t;
-                    var t = Bridge.Test.Runtime.TestFixture$1(Bridge.ClientTest.Batch4.SimpleTypes.DoubleTests).BeforeTest(true, assert, Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DoubleTests, void 0, ($t = new Bridge.Test.Runtime.TestContext(), $t.Method = "IFormattableToStringWorks()", $t.Line = "87", $t));
-                    t.Fixture.IFormattableToStringWorks();
-                },
-                ToStringWorks: function (assert) {
-                    var $t;
-                    var t = Bridge.Test.Runtime.TestFixture$1(Bridge.ClientTest.Batch4.SimpleTypes.DoubleTests).BeforeTest(true, assert, Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DoubleTests, void 0, ($t = new Bridge.Test.Runtime.TestContext(), $t.Method = "ToStringWorks()", $t.Line = "100", $t));
-                    t.Fixture.ToStringWorks();
-                },
-                ToExponentialWorks: function (assert) {
-                    var $t;
-                    var t = Bridge.Test.Runtime.TestFixture$1(Bridge.ClientTest.Batch4.SimpleTypes.DoubleTests).BeforeTest(true, assert, Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DoubleTests, void 0, ($t = new Bridge.Test.Runtime.TestContext(), $t.Method = "ToExponentialWorks()", $t.Line = "106", $t));
-                    t.Fixture.ToExponentialWorks();
-                },
-                ToExponentialWithFractionalDigitsWorks: function (assert) {
-                    var $t;
-                    var t = Bridge.Test.Runtime.TestFixture$1(Bridge.ClientTest.Batch4.SimpleTypes.DoubleTests).BeforeTest(true, assert, Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DoubleTests, void 0, ($t = new Bridge.Test.Runtime.TestContext(), $t.Method = "ToExponentialWithFractionalDigitsWorks()", $t.Line = "112", $t));
-                    t.Fixture.ToExponentialWithFractionalDigitsWorks();
-                },
-                ToFixed: function (assert) {
-                    var $t;
-                    var t = Bridge.Test.Runtime.TestFixture$1(Bridge.ClientTest.Batch4.SimpleTypes.DoubleTests).BeforeTest(true, assert, Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DoubleTests, void 0, ($t = new Bridge.Test.Runtime.TestContext(), $t.Method = "ToFixed()", $t.Line = "118", $t));
-                    t.Fixture.ToFixed();
-                },
-                ToFixedWithFractionalDigitsWorks: function (assert) {
-                    var $t;
-                    var t = Bridge.Test.Runtime.TestFixture$1(Bridge.ClientTest.Batch4.SimpleTypes.DoubleTests).BeforeTest(true, assert, Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DoubleTests, void 0, ($t = new Bridge.Test.Runtime.TestContext(), $t.Method = "ToFixedWithFractionalDigitsWorks()", $t.Line = "124", $t));
-                    t.Fixture.ToFixedWithFractionalDigitsWorks();
-                },
-                ToPrecisionWorks: function (assert) {
-                    var $t;
-                    var t = Bridge.Test.Runtime.TestFixture$1(Bridge.ClientTest.Batch4.SimpleTypes.DoubleTests).BeforeTest(true, assert, Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DoubleTests, void 0, ($t = new Bridge.Test.Runtime.TestContext(), $t.Method = "ToPrecisionWorks()", $t.Line = "130", $t));
-                    t.Fixture.ToPrecisionWorks();
-                },
-                ToPrecisionWithPrecisionWorks: function (assert) {
-                    var $t;
-                    var t = Bridge.Test.Runtime.TestFixture$1(Bridge.ClientTest.Batch4.SimpleTypes.DoubleTests).BeforeTest(true, assert, Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DoubleTests, void 0, ($t = new Bridge.Test.Runtime.TestContext(), $t.Method = "ToPrecisionWithPrecisionWorks()", $t.Line = "136", $t));
-                    t.Fixture.ToPrecisionWithPrecisionWorks();
-                },
-                IsPositiveInfinityWorks_SPI_1600: function (assert) {
-                    var $t;
-                    var t = Bridge.Test.Runtime.TestFixture$1(Bridge.ClientTest.Batch4.SimpleTypes.DoubleTests).BeforeTest(true, assert, Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DoubleTests, void 0, ($t = new Bridge.Test.Runtime.TestContext(), $t.Method = "IsPositiveInfinityWorks_SPI_1600()", $t.Line = "142", $t));
-                    t.Fixture.IsPositiveInfinityWorks_SPI_1600();
-                },
-                IsNegativeInfinityWorks: function (assert) {
-                    var $t;
-                    var t = Bridge.Test.Runtime.TestFixture$1(Bridge.ClientTest.Batch4.SimpleTypes.DoubleTests).BeforeTest(true, assert, Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DoubleTests, void 0, ($t = new Bridge.Test.Runtime.TestContext(), $t.Method = "IsNegativeInfinityWorks()", $t.Line = "154", $t));
-                    t.Fixture.IsNegativeInfinityWorks();
-                },
-                IsInfinityWorks: function (assert) {
-                    var $t;
-                    var t = Bridge.Test.Runtime.TestFixture$1(Bridge.ClientTest.Batch4.SimpleTypes.DoubleTests).BeforeTest(true, assert, Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DoubleTests, void 0, ($t = new Bridge.Test.Runtime.TestContext(), $t.Method = "IsInfinityWorks()", $t.Line = "164", $t));
-                    t.Fixture.IsInfinityWorks();
-                },
-                IsFiniteWorks: function (assert) {
-                    var $t;
-                    var t = Bridge.Test.Runtime.TestFixture$1(Bridge.ClientTest.Batch4.SimpleTypes.DoubleTests).BeforeTest(true, assert, Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DoubleTests, void 0, ($t = new Bridge.Test.Runtime.TestContext(), $t.Method = "IsFiniteWorks()", $t.Line = "174", $t));
-                    t.Fixture.IsFiniteWorks();
-                },
-                IsNaNWorks: function (assert) {
-                    var $t;
-                    var t = Bridge.Test.Runtime.TestFixture$1(Bridge.ClientTest.Batch4.SimpleTypes.DoubleTests).BeforeTest(true, assert, Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DoubleTests, void 0, ($t = new Bridge.Test.Runtime.TestContext(), $t.Method = "IsNaNWorks()", $t.Line = "183", $t));
-                    t.Fixture.IsNaNWorks();
-                },
-                GetHashCodeWorks: function (assert) {
-                    var $t;
-                    var t = Bridge.Test.Runtime.TestFixture$1(Bridge.ClientTest.Batch4.SimpleTypes.DoubleTests).BeforeTest(true, assert, Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DoubleTests, void 0, ($t = new Bridge.Test.Runtime.TestContext(), $t.Method = "GetHashCodeWorks()", $t.Line = "192", $t));
-                    t.Fixture.GetHashCodeWorks();
-                },
-                EqualsWorks: function (assert) {
-                    var $t;
-                    var t = Bridge.Test.Runtime.TestFixture$1(Bridge.ClientTest.Batch4.SimpleTypes.DoubleTests).BeforeTest(true, assert, Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DoubleTests, void 0, ($t = new Bridge.Test.Runtime.TestContext(), $t.Method = "EqualsWorks()", $t.Line = "201", $t));
-                    t.Fixture.EqualsWorks();
-                },
-                IEquatableEqualsWorks: function (assert) {
-                    var $t;
-                    var t = Bridge.Test.Runtime.TestFixture$1(Bridge.ClientTest.Batch4.SimpleTypes.DoubleTests).BeforeTest(true, assert, Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DoubleTests, void 0, ($t = new Bridge.Test.Runtime.TestContext(), $t.Method = "IEquatableEqualsWorks()", $t.Line = "210", $t));
-                    t.Fixture.IEquatableEqualsWorks();
-                },
-                CompareToWorks: function (assert) {
-                    var $t;
-                    var t = Bridge.Test.Runtime.TestFixture$1(Bridge.ClientTest.Batch4.SimpleTypes.DoubleTests).BeforeTest(true, assert, Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DoubleTests, void 0, ($t = new Bridge.Test.Runtime.TestContext(), $t.Method = "CompareToWorks()", $t.Line = "224", $t));
-                    t.Fixture.CompareToWorks();
-                },
-                IComparableCompareToWorks: function (assert) {
-                    var $t;
-                    var t = Bridge.Test.Runtime.TestFixture$1(Bridge.ClientTest.Batch4.SimpleTypes.DoubleTests).BeforeTest(true, assert, Bridge.Test.Runtime.BridgeClientTestBatch4Runner.DoubleTests, void 0, ($t = new Bridge.Test.Runtime.TestContext(), $t.Method = "IComparableCompareToWorks()", $t.Line = "233", $t));
-                    t.Fixture.IComparableCompareToWorks();
-                }
-            }
-        },
-        fields: {
-            context: null
-        },
-        methods: {
-            GetContext: function () {
-                var $t;
-                if (this.context == null) {
-                    this.context = ($t = new Bridge.Test.Runtime.FixtureContext(), $t.Project = "Batch4", $t.ClassName = "Bridge.ClientTest.Batch4.SimpleTypes.DoubleTests", $t.File = "Batch4\\SimpleTypes\\DoubleTests.cs", $t);
                 }
                 return this.context;
             }
