+++ Tests/Runner/Batch4/code.js
@@ -2186,156 +2186,6 @@ Bridge.assembly("Bridge.ClientTest.Batch4", function ($asm, globals) {
         }
     });
 
-    Bridge.define("Bridge.ClientTest.Batch4.SimpleTypes.DoubleTests", {
-        methods: {
-            TypePropertiesAreCorrect_SPI_1717: function () {
-                Bridge.Test.NUnit.Assert.True(Bridge.is(Bridge.box(0.5, System.Double, System.Double.format, System.Double.getHashCode), System.Double));
-                Bridge.Test.NUnit.Assert.AreEqual("System.Double", Bridge.Reflection.getTypeFullName(System.Double));
-                Bridge.Test.NUnit.Assert.False(Bridge.Reflection.isClass(System.Double));
-                Bridge.Test.NUnit.Assert.True(Bridge.Reflection.isAssignableFrom(System.IComparable$1(System.Double), System.Double));
-                Bridge.Test.NUnit.Assert.True(Bridge.Reflection.isAssignableFrom(System.IEquatable$1(System.Double), System.Double));
-                Bridge.Test.NUnit.Assert.True(Bridge.Reflection.isAssignableFrom(System.IFormattable, System.Double));
-                var d = Bridge.box(0.0, System.Double, System.Double.format, System.Double.getHashCode);
-                Bridge.Test.NUnit.Assert.True(Bridge.is(d, System.Double));
-                Bridge.Test.NUnit.Assert.True(Bridge.is(d, System.IComparable$1(System.Double)));
-                Bridge.Test.NUnit.Assert.True(Bridge.is(d, System.IEquatable$1(System.Double)));
-                Bridge.Test.NUnit.Assert.True(Bridge.is(d, System.IFormattable));
-
-                var interfaces = Bridge.Reflection.getInterfaces(System.Double);
-                Bridge.Test.NUnit.Assert.AreEqual(4, interfaces.length);
-                Bridge.Test.NUnit.Assert.True(System.Array.contains(interfaces, System.IComparable$1(System.Double), Function));
-                Bridge.Test.NUnit.Assert.True(System.Array.contains(interfaces, System.IEquatable$1(System.Double), Function));
-                Bridge.Test.NUnit.Assert.True(System.Array.contains(interfaces, System.IFormattable, Function));
-            },
-            GetDefaultValue: function (T) {
-                return Bridge.getDefaultValue(T);
-            },
-            DefaultValueIs0: function () {
-                Bridge.Test.NUnit.Assert.AreStrictEqual(0, this.GetDefaultValue(System.Double));
-            },
-            CreatingInstanceReturnsZero: function () {
-                Bridge.Test.NUnit.Assert.AreStrictEqual(0, Bridge.createInstance(System.Double));
-            },
-            ConstantsWork: function () {
-                var zero = 0;
-                Bridge.Test.NUnit.Assert.True(System.Double.min < System.Nullable.getValue(Bridge.cast(Bridge.unbox(Bridge.box(-1.7E+308, System.Double, System.Double.format, System.Double.getHashCode)), System.Double)), "MinValue should be correct");
-                Bridge.Test.NUnit.Assert.True(System.Double.max > System.Nullable.getValue(Bridge.cast(Bridge.unbox(Bridge.box(1.7E+308, System.Double, System.Double.format, System.Double.getHashCode)), System.Double)), "MaxValue should be correct");
-                // Not C# API
-                //Assert.AreEqual(double.JsMinValue, 5e-324, "MinValue should be correct");
-                Bridge.Test.NUnit.Assert.AreEqual(4.94065645841247E-324, 4.94065645841247E-324, "MinValue should be correct");
-                Bridge.Test.NUnit.Assert.True(isNaN(Number.NaN), "NaN should be correct");
-                Bridge.Test.NUnit.Assert.AreStrictEqual(1 / zero, Number.POSITIVE_INFINITY, "PositiveInfinity should be correct");
-                Bridge.Test.NUnit.Assert.AreStrictEqual(-1 / zero, Number.NEGATIVE_INFINITY, "NegativeInfinity should be correct");
-            },
-            DefaultConstructorReturnsZero: function () {
-                Bridge.Test.NUnit.Assert.AreStrictEqual(0, Number());
-            },
-            FormatWorks: function () {
-                Bridge.Test.NUnit.Assert.AreEqual("123", System.Double.format((291.0), "x"));
-            },
-            ToStringWithFormatWorks: function () {
-                Bridge.Test.NUnit.Assert.AreEqual("123", System.Double.format((291.0), "x"));
-            },
-            ToStringWithFormatAndProviderWorks: function () {
-                Bridge.Test.NUnit.Assert.AreEqual("123", System.Double.format((291.0), "x", System.Globalization.CultureInfo.invariantCulture));
-            },
-            IFormattableToStringWorks: function () {
-                Bridge.Test.NUnit.Assert.AreEqual("123", Bridge.format(291.0, "x", System.Globalization.CultureInfo.invariantCulture));
-            },
-            ToStringWorks: function () {
-                Bridge.Test.NUnit.Assert.AreEqual("123", System.Double.format((123.0)));
-            },
-            ToExponentialWorks: function () {
-                Bridge.Test.NUnit.Assert.AreEqual("1.23e+2", (123.0).toExponential());
-            },
-            ToExponentialWithFractionalDigitsWorks: function () {
-                Bridge.Test.NUnit.Assert.AreEqual("1.2e+2", (123.0).toExponential(1));
-            },
-            ToFixed: function () {
-                Bridge.Test.NUnit.Assert.AreEqual("123", (123.0).toFixed());
-            },
-            ToFixedWithFractionalDigitsWorks: function () {
-                Bridge.Test.NUnit.Assert.AreEqual("123.0", (123.0).toFixed(1));
-            },
-            ToPrecisionWorks: function () {
-                Bridge.Test.NUnit.Assert.AreEqual("12345", (12345.0).toPrecision());
-            },
-            ToPrecisionWithPrecisionWorks: function () {
-                Bridge.Test.NUnit.Assert.AreEqual("1.2e+4", (12345.0).toPrecision(2));
-            },
-            IsPositiveInfinityWorks_SPI_1600: function () {
-                var inf = Infinity;
-
-                // #1600
-                Bridge.Test.NUnit.Assert.True((inf === Number.POSITIVE_INFINITY));
-                Bridge.Test.NUnit.Assert.False((-inf === Number.POSITIVE_INFINITY));
-                Bridge.Test.NUnit.Assert.False((0.0 === Number.POSITIVE_INFINITY));
-                Bridge.Test.NUnit.Assert.False((Number.NaN === Number.POSITIVE_INFINITY));
-            },
-            IsNegativeInfinityWorks: function () {
-                var inf = Infinity;
-                Bridge.Test.NUnit.Assert.False((inf === Number.NEGATIVE_INFINITY));
-                Bridge.Test.NUnit.Assert.True((-inf === Number.NEGATIVE_INFINITY));
-                Bridge.Test.NUnit.Assert.False((0.0 === Number.NEGATIVE_INFINITY));
-                Bridge.Test.NUnit.Assert.False((Number.NaN === Number.NEGATIVE_INFINITY));
-            },
-            IsInfinityWorks: function () {
-                var inf = Infinity;
-                Bridge.Test.NUnit.Assert.True((Math.abs(inf) === Number.POSITIVE_INFINITY));
-                Bridge.Test.NUnit.Assert.True((Math.abs(-inf) === Number.POSITIVE_INFINITY));
-                Bridge.Test.NUnit.Assert.False((Math.abs(0.0) === Number.POSITIVE_INFINITY));
-                Bridge.Test.NUnit.Assert.False((Math.abs(Number.NaN) === Number.POSITIVE_INFINITY));
-            },
-            IsFiniteWorks: function () {
-                var zero = 0, one = 1;
-                Bridge.Test.NUnit.Assert.True(isFinite(one));
-                Bridge.Test.NUnit.Assert.False(isFinite(one / zero));
-                Bridge.Test.NUnit.Assert.False(isFinite(zero / zero));
-            },
-            IsNaNWorks: function () {
-                var zero = 0, one = 1;
-                Bridge.Test.NUnit.Assert.False(isNaN(one));
-                Bridge.Test.NUnit.Assert.False(isNaN(one / zero));
-                Bridge.Test.NUnit.Assert.True(isNaN(zero / zero));
-            },
-            GetHashCodeWorks: function () {
-                Bridge.Test.NUnit.Assert.AreEqual(System.Double.getHashCode((0.0)), System.Double.getHashCode((0.0)));
-                Bridge.Test.NUnit.Assert.AreEqual(System.Double.getHashCode((1.0)), System.Double.getHashCode((1.0)));
-                Bridge.Test.NUnit.Assert.AreNotEqual(System.Double.getHashCode((1.0)), System.Double.getHashCode((0.0)));
-                Bridge.Test.NUnit.Assert.AreNotEqual(System.Double.getHashCode((0.5)), System.Double.getHashCode((0.0)));
-            },
-            EqualsWorks: function () {
-                Bridge.Test.NUnit.Assert.True(System.Double.equals((0.0), Bridge.box(0.0, System.Double, System.Double.format, System.Double.getHashCode)));
-                Bridge.Test.NUnit.Assert.False(System.Double.equals((1.0), Bridge.box(0.0, System.Double, System.Double.format, System.Double.getHashCode)));
-                Bridge.Test.NUnit.Assert.False(System.Double.equals((0.0), Bridge.box(0.5, System.Double, System.Double.format, System.Double.getHashCode)));
-                Bridge.Test.NUnit.Assert.True(System.Double.equals((1.0), Bridge.box(1.0, System.Double, System.Double.format, System.Double.getHashCode)));
-            },
-            IEquatableEqualsWorks: function () {
-                Bridge.Test.NUnit.Assert.True((0.0) === 0.0);
-                Bridge.Test.NUnit.Assert.False((1.0) === 0.0);
-                Bridge.Test.NUnit.Assert.False((0.0) === 0.5);
-                Bridge.Test.NUnit.Assert.True((1.0) === 1.0);
-
-                Bridge.Test.NUnit.Assert.True(Bridge.equalsT(0.0, 0.0, System.Double));
-                Bridge.Test.NUnit.Assert.False(Bridge.equalsT(1.0, 0.0, System.Double));
-                Bridge.Test.NUnit.Assert.False(Bridge.equalsT(0.0, 0.5, System.Double));
-                Bridge.Test.NUnit.Assert.True(Bridge.equalsT(1.0, 1.0, System.Double));
-            },
-            CompareToWorks: function () {
-                Bridge.Test.NUnit.Assert.True(Bridge.compare((0.0), 0.0) === 0);
-                Bridge.Test.NUnit.Assert.True(Bridge.compare((1.0), 0.0) > 0);
-                Bridge.Test.NUnit.Assert.True(Bridge.compare((0.0), 0.5) < 0);
-                Bridge.Test.NUnit.Assert.True(Bridge.compare((1.0), 1.0) === 0);
-            },
-            IComparableCompareToWorks: function () {
-                Bridge.Test.NUnit.Assert.True(Bridge.compare(0.0, 0.0, false, System.Double) === 0);
-                Bridge.Test.NUnit.Assert.True(Bridge.compare(1.0, 0.0, false, System.Double) > 0);
-                Bridge.Test.NUnit.Assert.True(Bridge.compare(0.0, 0.5, false, System.Double) < 0);
-                Bridge.Test.NUnit.Assert.True(Bridge.compare(1.0, 1.0, false, System.Double) === 0);
-            }
-        }
-    });
-
     Bridge.define("Bridge.ClientTest.Batch4.SimpleTypes.Int16Tests", {
         methods: {
             TypePropertiesAreCorrect_SPI_1717: function () {
